requirements.txt
python-can~=3.3.4
bitstruct~=8.11.1
argparse-addons~=0.7.0
